refactor: ChatPanel - gql 쿼리 분리
